+++ CHANGELOG.md
@@ -5,10 +5,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [1.1.0] - 2022-05-05
-### Fixed
-- `Input`: handle label float when user set value via `useForm`'s `setValue` (#23)
+## [1.1.0] - 2022-05-08
 
+### Added
+
+- DateTimePicker
+- MaskedInput (no stories yet)
+- TimePicker (unfinished - unstable)
+- useMediaQuery hook
+- Add `dd-` prefix to components
+
+### Fixed
+
+- `Input`: handle label float when user set value via `useForm`'s `setValue` (#23)
+- `Button`: fix issue with no content on `isLoading` `true`
 
 ## [1.0.14] - 2022-04-29